% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace_clear_selection.R
\name{ace_clear_selection}
\alias{ace_clear_selection}
\title{Clear the editor selected range.}
\usage{
ace_clear_selection(id, session = getDefaultReactiveDomain(), as_is = FALSE)
}
\arguments{
\item{id}{\code{[string]}}

\item{session}{\code{[NULL]}  Defaults to \code{getDefaultReactiveDomain()}.}

\item{as_is}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{FALSE}}
}
\value{
\code{ace_clear_selection}: invisible(NULL)
}
\description{
Clear the editor selected range.
}
