% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_editable_class.R
\name{add_editable_class}
\alias{add_editable_class}
\title{Add an editable class.}
\usage{
add_editable_class(inputId, editable_class, assert_unique = TRUE)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{editable_class}{\code{[string]}}

\item{assert_unique}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{TRUE}}
}
\value{
\code{add_editable_class}: \code{[html]}
}
\description{
Add an editable class.
}
\examples{
 add_editable_class(inputId = 'myid', editable_class = 'pill-item',
 assert_unique = TRUE)
 add_editable_class(inputId = 'myid', editable_class = 'pill-item',
 assert_unique = FALSE)
}
