% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_sheet.R
\name{ao_sheet}
\alias{ao_sheet}
\title{create an AO sheet binding.}
\usage{
ao_sheet(..., inputId)
}
\arguments{
\item{inputId}{\code{[string]}}
}
\value{
\code{ao_sheet}: \code{[HTML]}
}
\description{
create an AO sheet binding.
}
\examples{
 ao_sheet(inputId = 'myid')
}
