% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_class.R
\name{append_class}
\alias{append_class}
\title{Append the class of a shiny tag.}
\usage{
append_class(x, class)
}
\arguments{
\item{x}{[class]}
}
\value{
\code{append_class}: shiny tag
}
\description{
Append the class of a shiny tag.
}
\examples{
 x <- div(class = 'class1')
 append_class(x) <- 'new_class'
 x
}
