% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_shiny_button.R
\name{as_shiny_button}
\alias{as_shiny_button}
\title{Convert an html element into a shiny action button.}
\usage{
as_shiny_button(tag)
}
\arguments{
\item{tag}{[class]}
}
\value{
\code{as_shiny_button}: HTML (shiny.tag)
}
\description{
Convert an html element into a shiny action button.
}
\examples{
 if (interactive()) { 
 library('shiny') 
 library('ShinyReboot') 
 bs_global_theme() 
 ui <- fluidPage(bs_dependencies(theme = bs_global_get()), 
 tags$head(reboot_Dependancies()), tags$h1('as_shiny_button'), 
 as_shiny_button(div(id = 'shiny_button', class = 'border', 
 'click me!')), verbatimTextOutput(outputId = 'res')) 
 server <- function(input, output, session) { 
 output$res <- renderPrint(input[['shiny_button']]) 
 } 
 shinyApp(ui, server) 
 } 
}
