% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_css_unit.R
\name{assert_css_unit}
\alias{assert_css_unit}
\title{Checks that the argument is valid for use as a CSS unit of length.}
\usage{
assert_css_unit(x)
}
\arguments{
\item{x}{\code{[number|string]}}
}
\value{
\code{assert_css_unit}: \code{[invisible(x)]} if valid, otherwise throws an error.
}
\description{
Checks that the argument is valid for use as a CSS unit of length.
}
