% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox_group_input.R
\name{checkbox_group_input}
\alias{checkbox_group_input}
\title{Create a radio slider input.}
\usage{
checkbox_group_input(
  ...,
  inputId,
  label,
  choices,
  selected = NULL,
  clear_filter = TRUE,
  search = TRUE,
  options_dropdown = TRUE
)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{label}{\code{[string]}  NULL is ok.}

\item{choices}{\code{[atomic_vector]}}

\item{selected}{\code{[choice]}  Possible values: \code{choices}.  NULL is ok.  Defaults to \code{NULL}}

\item{clear_filter}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{TRUE}}

\item{search}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{TRUE}}

\item{options_dropdown}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{TRUE}}
}
\value{
\code{checkbox_group_input}: \code{[html]}
}
\description{
Create a radio slider input.
}
\examples{
 checkbox_group_input(inputId = 'myid', label = 'label', choices = names(mtcars))
}
