% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_picker_btn.R
\name{color_picker_btn}
\alias{color_picker_btn}
\title{Generate a color picker button.}
\usage{
color_picker_btn(inputId, color, ...)
}
\arguments{
\item{inputId}{[string]}

\item{color}{[string]   All non-missing elements must comply to regex pattern ^#.  Defaults to '#ffffff'}
}
\value{
\code{color_picker_btn}: HTML
}
\description{
Generate a color picker button.
}
\examples{
 if (interactive()) {
 library('shiny')
 library('ShinyReboot')
 ui <- fluidPage(tagList(html_dependency_shinyReboot()),
 tags$h1('Spectrum color picker'), uiOutput('ui_colorPicker'),
 verbatimTextOutput(outputId = 'res'))
 server <- function(input, output, session) {
 output$res <- renderPrint(input$myColor)
 output$ui_colorPicker <- renderUI({
 color_picker_btn('myColor')
 })
 }
 shinyApp(ui, server)
 }
}
