% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_seq.R
\name{color_seq}
\alias{color_seq}
\title{Create a color sequence.}
\usage{
color_seq(colors, from, to, length)
}
\arguments{
\item{colors}{\code{[character]}}

\item{from}{\code{[number]}  Must be greater than \code{0}.  Must be less than \code{1}.}

\item{to}{\code{[number]}  Must be greater than \code{from}.  Must be less than \code{1}.}

\item{length}{\code{[number]}}
}
\value{
\code{color_seq}: \code{[character]} A character vector of hex colors
}
\description{
Create a color sequence.
}
\examples{
 color_seq(colors = c('white', 'purple'), from = 0.25, to = 1,
 length = 3)
}
