% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_palette_ui.R
\name{continuous_palette_ui}
\alias{continuous_palette_ui}
\title{Render a continuous palette ui.}
\usage{
continuous_palette_ui(
  inputId,
  data_range,
  selected = "Blues",
  stepped = FALSE,
  n_steps = 4,
  selected_range = NULL,
  selected_values = NULL,
  color_min = NULL,
  color_max = NULL
)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{data_range}{\code{[numeric]}  Must have an exact length of \code{2}.}

\item{selected}{\code{[string]}  Defaults to \code{'Blues'}}
}
\value{
\code{continuous_palette_ui}: \code{[html]}
}
\description{
Render a continuous palette ui.
}
