% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_pallete_picker.R
\name{continuous_pallete_picker}
\alias{continuous_pallete_picker}
\title{create a continuous color palette picker.}
\usage{
continuous_pallete_picker(inputId, package = "grDevices", selected = "Blues")
}
\arguments{
\item{inputId}{\code{[string]}}

\item{package}{\code{[subset]}  Defaults to \code{'grDevices'}}

\item{selected}{\code{[choice]}  Possible values: \code{palette_names}.  Defaults to \code{NULL}}
}
\value{
\code{continuous_pallete_picker}: \code{[html]}
}
\description{
create a continuous color palette picker.
}
\examples{
 continuous_pallete_picker(inputId = 'myid')
}
