% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ace_token.R
\name{create_ace_token}
\alias{create_ace_token}
\title{create an ace token.}
\usage{
create_ace_token(name, values, escape = TRUE)
}
\arguments{
\item{name}{\code{[string]}}

\item{values}{\code{[character]}}

\item{escape}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{TRUE}}
}
\value{
\code{create_ace_token}: \code{[list(token=string(),regex=string())]}
}
\description{
create an ace token.
}
\examples{
 create_ace_token(name = 'variables', values = c('V(1)', 'V(2)',
 'V(3)'), escape = TRUE)
 create_ace_token(name = 'functions', values = paste0(c('SUM',
 'MIN', 'MAX'), followed_by('\\\\(')), escape = FALSE)
}
