% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dropdown.R, R/dropdown_ui.R
\name{data_dropdown}
\alias{data_dropdown}
\alias{dropdown_ui}
\title{Create a data dropdown.}
\usage{
data_dropdown(
  ...,
  inputId,
  target,
  handle_class = NULL,
  class = NULL,
  width = "auto",
  position = "absolute"
)

dropdown_ui(
  ...,
  inputId,
  target,
  handle_class = NULL,
  class = NULL,
  width = NULL
)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{handle_class}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}

\item{class}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}

\item{target_class}{\code{[string]}}
}
\value{
\code{data_dropdown}: \code{[html]}

\code{dropdown_ui}: \code{[html]}
}
\description{
Create a data dropdown.

Create a dropdown ui
}
\examples{
 data_dropdown(dropdown_button(inputId = 'edit', label = 'Edit...'),
 dropdown_checkbox(inputId = 'filter', value = FALSE,
 label = 'Show Filter'), dropdown_radio_group(inputId = 'radio_nformat',
 options = c('currency', 'percentage', 'float', 'integer'),
 labels = c('currency', 'percentage', 'float', 'integer')),
 dropdown_submenu(label = 'Units', dropdown_radio_group(inputId = 'radio_units',
 labels = c('duration', 'distance', 'mass', 'speed'),
 options = c('duration', 'distance', 'mass', 'speed'))),
 inputId = 'myid', handle_class = 'dropdown-handle', target_class = 'pill-item')

 if(interactive()){
 library("shiny")
 library("ShinyReboot")
 library("bslib")
 bs_global_theme()
 ui <- fluidPage(
 bs_theme_dependencies(theme = bs_global_get()),

 tags$h1("Data DropDown"),
 icon_mdi('pencil'),
 flexCol(class='w-50',
 pill_card(inputId ="nav_dimensions",
 type='column',
 sort_ops = sortable_options(
 name = 'nav_dimensions',
 pull = 'clone',
 put = FALSE,
 dragClass = 'drag',
 sort = FALSE
 ),
 pill_item(
 id = c('sales1', 'transactions1', 'traffic1',
 'basket_size1'),
 label = c('sales', 'transactions', 'traffic',
 'basket size'),
 data=list(
 nformat=c('currency','integer','integer',NA),
 filter=c(TRUE,FALSE,NA,FALSE),
 edit=c(NA,0,0,0)
 )
 )),
 data_dropdown(dropdown_button(inputId='action_edit',label='Edit...'),
 dropdown_checkbox(inputId='filter',value=FALSE,label='Show Filter'),
 dropdown_radio_group(inputId = 'radio_nformat',
 options = c('currency',
 'percentage',
 'float',
 'integer'
 ),
 labels = c('currency',
 'percentage',
 'float',
 'integer'
 )),
 dropdown_submenu(label = 'Units',
 dropdown_radio_group(
 inputId = 'radio_units',
 labels = c('duration','distance','mass','speed'),
 options= c('duration','distance','mass','speed'))) ,
 inputId='myid',
 handle_class = 'dropdown-handle',
 target='.pill-item')),
 verbatimTextOutput('res')

 )

 server = function(input, output, session) {

 output$res<-renderText({
 glue("{names(input$myid)}:{unlist(input$myid)}")\%sep\%'\n'
 })
 observe({
 print(input$myid)
 })
 observe({
 print(input$myid_edit)
 })
 }
 shinyApp(ui = ui, server = server)
 }
}
