% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_shelf.R
\name{data_filters}
\alias{data_filters}
\title{Create a filter shelf binding.}
\usage{
data_filters(inputId, ...)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{...}{html attributes or children.  Filters should have the class 'data-filter'}
}
\value{
\code{data_filters}: \code{[html]}
}
\description{
Create a filter shelf binding.
}
\examples{
 data_filters(inputId = 'sd')
}
