% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_header.R
\name{dd_header}
\alias{dd_header}
\title{Create a drop down menu header.}
\usage{
dd_header(..., id, class = "", carat = TRUE, is_submenu = FALSE)
}
\arguments{
\item{id}{[character]  Must have an exact length of 1.}

\item{class}{[character]  Must have an exact length of 1.  Defaults to ''}

\item{carat}{[logical]  Must have an exact length of 1.  Defaults to TRUE}

\item{is_submenu}{[logical]  Defaults to FALSE}
}
\value{
\code{dd_header}: [HTML]
}
\description{
For use inside ml_dropdown
}
\examples{
 header = 'Drop Down'
 ml_dropdown(class = 'multi-level-dropdown', dd_header(header,
 id = label2id(header), class = 'menu-text', carat = FALSE),
 dd_menu(id = label2id(header), class = 'shadow', dd_item(id = 'option1',
 label = 'Option 1'), dd_item(id = 'option2', label = 'Option 2'),
 dd_item(id = 'option3', label = 'Option 3')))
}
