% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_item.R
\name{dd_item}
\alias{dd_item}
\title{Create a drop down menu item.}
\usage{
dd_item(..., id, label, class = "")
}
\arguments{
\item{id}{[character]  Must have an exact length of 1.}

\item{label}{[character]  Must have an exact length of 1.}

\item{class}{[character]  Must have an exact length of 1.  Defaults to ''}
}
\value{
\code{dd_item}: [HTML]
}
\description{
Create a drop down menu item.
}
\examples{
 dd_item(id = 'option2', label = 'Option 2', ) 
}
