% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_menu.R
\name{dd_menu}
\alias{dd_menu}
\title{Create a drop down menu list.}
\usage{
dd_menu(..., id, class = "")
}
\arguments{
\item{id}{[character]  Must have an exact length of 1.}

\item{class}{[character]  Must have an exact length of 1.  Defaults to ''}
}
\value{
\code{dd_menu}: [HTML]
}
\description{
For use inside ml_dropdown
}
\examples{
 header = 'Drop Down'
 ml_dropdown(class = 'multi-level-dropdown', dd_header(header,
 id = label2id(header), class = 'menu-text', carat = FALSE),
 dd_menu(id = label2id(header), class = 'shadow', dd_item(id = 'option1',
 label = 'Option 1'), dd_item(id = 'option2', label = 'Option 2'),
 dd_item(id = 'option3', label = 'Option 3')))
}
