% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_color_picker.R
\name{default_color_picker}
\alias{default_color_picker}
\title{Generate a default color picker.}
\usage{
default_color_picker(
  inputId,
  value = "#ffffff",
  width = "25px",
  height = "25px",
  ...
)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{value}{\code{[string]}  All non-missing elements must comply to regex pattern \code{'''^#'''}.  NULL is ok.  Defaults to \code{'#ffffff'}}

\item{height}{\code{[string]}  All non-missing elements must comply to regex pattern \code{'''\\d+px$'''}.  Defaults to \code{'25px'}}
}
\value{
\code{default_color_picker}: \code{[html]}
}
\description{
Generate a default color picker.
}
