% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_pallete_picker.R
\name{discrete_pallete_picker}
\alias{discrete_pallete_picker}
\title{create a discrete color palette picker.}
\usage{
discrete_pallete_picker(
  inputId,
  package = "ggthemes",
  min_colors = 7,
  selected = NULL
)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{package}{\code{[subset]}  Defaults to \code{'ggthemes'}}

\item{min_colors}{\code{[integerish]}  Defaults to \code{7}}

\item{selected}{\code{[choice]}  Possible values: \code{palette_names}.  Defaults to \code{NULL}}
}
\value{
\code{discrete_pallete_picker}: \code{[html]}
}
\description{
create a discrete color palette picker.
}
\examples{
 discrete_pallete_picker(inputId = 'myid')
}
