% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_variable_palette_picker.R
\name{discrete_variable_palette_picker}
\alias{discrete_variable_palette_picker}
\title{Create a discrete variable color palette selector.}
\usage{
discrete_variable_palette_picker(
  inputId,
  var_names,
  var_ids = NULL,
  colors = NULL,
  palette = NULL
)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{var_names}{\code{[character]}}

\item{var_ids}{\code{[character]}  NULL is ok.  Defaults to \code{NULL}}

\item{colors}{\code{[character]}  NULL is ok.  Must have an exact length of \code{length(var_names)}.  Defaults to \code{NULL}}

\item{palette}{\code{[choice]}  Possible values: \code{ggthemes_palette_names()}.  NULL is ok.  Defaults to \code{NULL}}
}
\value{
\code{discrete_variable_palette_picker}: \code{[html]}
}
\description{
Create a discrete variable color palette selector.
}
