% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draggable_model.R
\name{draggable_model}
\alias{draggable_model}
\title{Create a draggable model.}
\usage{
draggable_model(inputId, left, top, width = NULL, ...)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{left}{\code{[number]}  Must be greater than \code{0}.  Must be less than \code{100}.}

\item{top}{\code{[number]}  Must be greater than \code{0}.  Must be less than \code{100}.}
}
\value{
\code{draggable_model}: \code{[html]}
}
\description{
Create a draggable model.
}
\examples{
 draggable_model(inputId = 'drag', left = 10, top = 10, width = '30\%')
 draggable_model(inputId = 'drag', class = 'myModel', left = 10,
 top = 10, width = '30\%', div('someContent'))
 draggable_model(inputId = 'drag', left = 100, top = 10, width = '30px')
 draggable_model(inputId = 'drag', left = 120, top = 10, width = '30\%')
}
