% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draggable_model_btn.R
\name{draggable_model_btn}
\alias{draggable_model_btn}
\title{Create a button to show or hide a draggable model.}
\usage{
draggable_model_btn(..., inputId, toggleId, outerTag = "button")
}
\arguments{
\item{inputId}{\code{[string]}}

\item{toggleId}{\code{[string]}}

\item{outerTag}{\code{[string]}  Defaults to \code{'button'}}
}
\value{
\code{draggable_model_btn}: \code{[html]}
}
\description{
Create a button to show or hide a draggable model.
}
\examples{
 draggable_model_btn(inputId = 'drag_btn', toggleId = 'drag',
 'Show')
 draggable_model_btn(inputId = 'drag_btn', toggleId = 'drag',
 'Show', outerTag = 'div')
 if(interactive()) {
 library("shiny")
 library("ShinyReboot")
 library("bslib")
 bs_global_theme()
 ui <- fluidPage(
 bs_theme_dependencies(theme = bs_global_get()),

 tags$h1("Draggable Model"),
 draggable_model_btn(
 inputId = "show_model",
 toggleId = 'example',
 class = "btn btn-primary new-var-btn",
 "Launch demo modal"
 ),
 draggable_model(
 class = "card rounded-0 bg-light p-1",
 inputId = "example",
 top = 40,
 left = 20,
 div(class = 'drag-handle w-100', "Drag Me"),
 div(
 class = "d-flex mt-4",
 draggable_model_btn(
 inputId = "ok_close",
 toggleId = 'example',
 class = "btn btn-outline-dark btn-sm rounded-0 var-btn mr-1",
 'OK'
 ),
 tags$button(class = "btn btn-outline-dark btn-sm rounded-0 var-btn mr-auto",
 "Apply"),
 draggable_model_btn(
 inputId = "cancel_close",
 toggleId = 'example',
 class = "btn btn-outline-dark btn-sm rounded-0 var-btn",
 "Cancel"
 )
 )
 )
 )

 server = function(input, output, session) {
 observe({
 print(input$example)
 })
 }
 shinyApp(ui = ui, server = server)
 }
}
