% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown_button.R
\name{dropdown_button}
\alias{dropdown_button}
\title{Create a action button in a dropdown menu.}
\usage{
dropdown_button(inputId, label)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{label}{\code{[string]}}
}
\value{
\code{dropdown_button}: \code{[html]}
}
\description{
Create a action button in a dropdown menu.
}
\examples{
 dropdown_button(inputId = 'edit', label = 'Edit...')
}
