% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown_checkbox.R
\name{dropdown_checkbox}
\alias{dropdown_checkbox}
\title{Create a checkbox in a dropdown menu.}
\usage{
dropdown_checkbox(inputId, label, value = FALSE)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{value}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{FALSE}}
}
\value{
\code{dropdown_checkbox}: \code{[html]}
}
\description{
Create a checkbox in a dropdown menu.
}
\examples{
 dropdown_checkbox(inputId = 'filter', value = FALSE, label = 'Show Filter')
}
