% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown_menu.R
\name{dropdown_menu}
\alias{dropdown_menu}
\title{Create a drop down menu.}
\usage{
dropdown_menu(..., label, class = "")
}
\arguments{
\item{label}{[character]  Must have an exact length of 1.}

\item{class}{[character]  Must have an exact length of 1.  Defaults to ''}
}
\value{
\code{dropdown_menu}: HTML
}
\description{
Create a drop down menu.
}
\examples{
 dropdown_menu(label = 'File', class = 'menu-item', dropdown_btn_group(inputIds = letters[1:3],
 labels = c('Add to Sheet', 'New Calculation', 'Show Filter'),
 types = c('action', 'action', 'checkbox')), dropdown_btn_group(inputIds = letters[4:6],
 labels = c('Copy', 'Duplicate', 'Edit'), types = c('action',
 'action', 'model')), dropdown_radio_group(inputId = 'letters',
 options = letters[1:4], labels = c('Currency', 'Percentage',
 'Number', 'Units')), dropdown_submenu(label = 'More Options',
 dropdown_btn_group(inputIds = letters[1:2], labels = letters[1:2],
 types = c('action', 'action'))))
}
