% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown_radio_group.R
\name{dropdown_radio_group}
\alias{dropdown_radio_group}
\title{Create an drop down radio button group.}
\usage{
dropdown_radio_group(
  inputId,
  options,
  labels,
  class = "",
  check_icon = "checkbox-blank-circle"
)
}
\arguments{
\item{inputId}{[character]  Must have an exact length of 1.}

\item{labels}{[character]}

\item{class}{[character]  Defaults to ''}

\item{check_icon}{[character]  Must have an exact length of 1.  Defaults to 'checkbox-blank-circle'}
}
\value{
\code{dropdown_radio_group}: [HTML]
}
\description{
Create an drop down radio button group.
}
\examples{
 dropdown_radio_group(inputId = 'numberformat', options = letters[1:4],
 labels = c('Currency', 'Percentage', 'Number', 'Units'))
}
