% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown_submenu.R
\name{dropdown_submenu}
\alias{dropdown_submenu}
\title{Create a drop down submenu.}
\usage{
dropdown_submenu(..., label, parent = NULL)
}
\arguments{
\item{label}{[character]  Must have an exact length of 1.}

\item{parent}{[NULL]  Defaults to NULL}
}
\value{
\code{dropdown_submenu}: HTML
}
\description{
Create a drop down submenu.
}
\examples{
 dropdown_submenu(label = 'More Options', dropdown_btn_group(inputIds = letters[1:2], 
 labels = letters[1:2], types = c('action', 'action'))) 
}
