% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather_icon.R
\name{feather_icon}
\alias{feather_icon}
\title{Use a feather icon.}
\usage{
feather_icon(name, ...)
}
\arguments{
\item{name}{[character]  Name of Icon. See https://feathericons.com/. Must have an exact length of 1.}
}
\value{
\code{feather_icon}: html
}
\description{
Use a feather icon.
}
