% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_shelf.R
\name{filter_shelf}
\alias{filter_shelf}
\title{Create a filter shelf binding.}
\usage{
filter_shelf(inputId, ...)
}
\arguments{
\item{inputId}{\code{[string]}}
}
\value{
\code{filter_shelf}: \code{[html]}
}
\description{
Create a filter shelf binding.
}
\examples{
 filter_shelf(inputId = 'sd')
}
