% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexBox.R
\name{flexBox}
\alias{flexBox}
\title{Create a BS4 flex box.}
\usage{
flexBox(...)
}
\arguments{
\item{...}{html attributes and inner html}
}
\value{
\code{flexBox}: HTML
}
\description{
Create a BS4 flex box.
}
