% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexCol.R
\name{flexCol}
\alias{flexCol}
\title{Create a BS4 flex column.}
\usage{
flexCol(...)
}
\arguments{
\item{...}{html attributes and inner html}
}
\value{
\code{flexCol}: HTML
}
\description{
Create a BS4 flex column.
}
