% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexRow.R
\name{flexRow}
\alias{flexRow}
\title{Create a BS4 flex row.}
\usage{
flexRow(...)
}
\arguments{
\item{...}{html attributes and inner html}
}
\value{
\code{flexRow}: HTML
}
\description{
Create a BS4 flex row.
}
