% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_session_ns.R
\name{get_session_ns}
\alias{get_session_ns}
\title{Get a session namespace}
\usage{
get_session_ns(session = getDefaultReactiveDomain())
}
\arguments{
\item{session}{the shiny session.}
}
\value{
the session's namespace function
}
\description{
Get a session namespace
}
