% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_divs.R
\name{gradient_divs}
\alias{gradient_divs}
\title{Create a series of div colored by a stepped gradient.}
\usage{
gradient_divs(..., n, domain, range)
}
\value{
\code{gradient_divs}: \code{[list]}
}
\description{
Create a series of div colored by a stepped gradient.
}
\examples{
 gradient_divs(class = 'flex-fill', n = 5, domain = continuous_pal('Blues',
 n = 3), range = c(0, 4, 5))
}
