% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_scale.R
\name{gradient_scale}
\alias{gradient_scale}
\title{Create a number gradient interpolator.}
\usage{
gradient_scale(domain, range)
}
\arguments{
\item{domain}{\code{[character]}  All non-missing elements must comply to regex pattern \code{'''^#'''}.}

\item{range}{\code{[numeric]}  Must have an exact length of \code{length(domain)}.}
}
\value{
\code{gradient_scale}: \code{[function]}
}
\description{
Create a number gradient interpolator.
}
