% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasHTMLclass.R
\name{hasHTMLclass}
\alias{hasHTMLclass}
\title{Test is a shiny tag as a specific html class.}
\usage{
hasHTMLclass(x, class)
}
\arguments{
\item{x}{[class]}

\item{class}{[character]}
}
\value{
\code{hasHTMLclass}: Logical(1)
}
\description{
Test is a shiny tag as a specific html class.
}
\examples{
 x = div(class = 'class1 active show') 
 hasHTMLclass(x, 'active') 
 hasHTMLclass(x, c('active', 'show')) 
 hasHTMLclass(x, c('active', 'class2')) 
 hasHTMLclass(x, c('class1 show')) 
}
