% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_class.R
\name{html_class}
\alias{html_class}
\title{Get or set the class of a shiny tag.}
\usage{
html_class(x)
}
\arguments{
\item{x}{[class]}
}
\value{
\code{html_class}: character(1)
}
\description{
Get or set the class of a shiny tag.
}
\examples{
 x <- div()
 html_class(x)
 html_class(x) <- 'new_class'
 html_class(x)
}
