% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_id.R
\name{html_id}
\alias{html_id}
\title{Get or set the id of a shiny tag.}
\usage{
html_id(x)
}
\arguments{
\item{x}{[class]}
}
\value{
\code{html_id}: character(1)
}
\description{
Get or set the id of a shiny tag.
}
\examples{
 x <- div()
 html_id(x)
 html_id(x) <- 'ds'
 html_id(x)
}
