% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_btn.R
\name{icon_btn}
\alias{icon_btn}
\title{Create an icon button.}
\usage{
icon_btn(
  inputId,
  icon,
  toggleId = NULL,
  tooltip = NULL,
  btn_style = "light",
  size = "sm",
  ...
)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{icon}{\code{[string]}}

\item{toggleId}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}

\item{tooltip}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}

\item{btn_style}{\code{[string]}  Defaults to \code{'light'}}

\item{size}{\code{[choice]}  Defaults to \code{'sm'}}
}
\value{
\code{icon_btn}: \code{[html]}
}
\description{
Create an icon button.
}
\examples{
 icon_btn(inputId = 'hi', icon = 'pencil', toggleId = NULL)
 icon_btn(inputId = 'hi', icon = 'pencil', toggleId = 'model')
 icon_btn(inputId = 'hi', icon = 'pencil', toggleId = 'model',
 tooltip = 'launch model', size = 'md')
}
