% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_button_group.R
\name{icon_button_group}
\alias{icon_button_group}
\title{Create an icon button group.}
\usage{
icon_button_group(
  inputId,
  icon_name,
  tooltip,
  class = "",
  tooltip_placement = "top"
)
}
\arguments{
\item{inputId}{[character]}

\item{icon_name}{[character]  Must have an exact length of length.  Must have an exact length of inputId.}

\item{tooltip}{[character]  Must have an exact length of length.  Must have an exact length of inputId.}

\item{class}{[character]  Defaults to ''}

\item{tooltip_placement}{[character]  Must have an exact length of 1.  Defaults to 'top'}
}
\value{
\code{icon_button_group}: [HTML]
}
\description{
Create an icon button group.
}
\examples{
 icon_button_group(inputId = c('undo', 'redo', 'save'), icon_name = c('undo',
 'redo', 'content-save-outline'), tooltip = c('Undo',
 'Redo', 'Save'))
}
