% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_radio_group.R
\name{icon_radio_group}
\alias{icon_radio_group}
\title{Create an icon button group.}
\usage{
icon_radio_group(
  inputId,
  options,
  icon_name,
  tooltip,
  class = "",
  tooltip_placement = "top"
)
}
\arguments{
\item{inputId}{[character]  Must have an exact length of 1.}

\item{icon_name}{[character]}

\item{tooltip}{[character]  Must have an exact length of length.  Must have an exact length of icon_name.}

\item{class}{[character]  Defaults to ''}

\item{tooltip_placement}{[character]  Must have an exact length of 1.  Defaults to 'top'}
}
\value{
\code{icon_radio_group}: [HTML]
}
\description{
Create an icon button group.
}
\examples{
 icon_radio_group(inputId = 'sort_options', options = c('sort_remove',
 'sort_ascending', 'sort_descending'), icon_name = c('sort-variant-remove',
 'sort-ascending', 'sort-descending'), tooltip = c('Clear sort',
 'Sort ascending', 'Sort descending'))
}
