% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_tool_bar.R
\name{icon_tool_bar}
\alias{icon_tool_bar}
\title{Create an icon tool bar wrapper.}
\usage{
icon_tool_bar(..., class = "")
}
\arguments{
\item{class}{[character]  Must have an exact length of 1.  Defaults to ''}
}
\value{
\code{icon_tool_bar}: [HTML]
}
\description{
Create an icon tool bar wrapper.
}
\examples{
 icon_tool_bar(icon_button_group(inputId = c('undo', 'redo',
 'save'), c('undo', 'redo', 'content-save-outline'), c('Undo',
 'Redo', 'Save')), icon_button_group(c('data_source_new',
 'data_source_refresh', 'pause_data_updates'), c('database-plus',
 'database-refresh', 'database-lock'), c('New data source',
 'Refresh data source', 'Pause auto updates')), icon_radio_group(inputId = 'sort_options',
 c('sort_remove', 'sort_ascending', 'sort_descending'),
 c('sort-variant-remove', 'sort-ascending', 'sort-descending'),
 c('Clear sort', 'Sort ascending', 'Sort descending')))
}
