% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_logical.R
\name{js_logical}
\alias{js_logical}
\title{Convert R TRUE/FALSE to HTML/JS true/false.}
\usage{
js_logical(x)
}
\arguments{
\item{x}{[logical]}
}
\value{
\code{js_logical}: character
}
\description{
Convert R TRUE/FALSE to HTML/JS true/false.
}
\examples{
 js_logical(c(T, F)) 
 js_logical(c(TRUE, FALSE)) 
}
