% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label2id.R
\name{label2id}
\alias{label2id}
\title{Turn a html label into an shiny input id.}
\usage{
label2id(x)
}
\arguments{
\item{x}{[character]}
}
\value{
\code{label2id}: tolower(str_replace_all([character(1)],'\s','_')
}
\description{
Turn a html label into an shiny input id.
}
\examples{
 label2id('Open File') 
}
