% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_app_css.R
\name{load_app_css}
\alias{load_app_css}
\title{Load all CSS files in the www directory.}
\usage{
load_app_css()
}
\arguments{
\item{app_name}{[character]  NULL is ok.  Must have an exact length of 1.  Defaults to NULL}
}
\value{
\code{load_app_css}: html
}
\description{
Load all CSS files in the www directory.
}
