% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Create a div for main content.}
\usage{
main(..., class = "col-md-9 ml-sm-auto col-lg-10 p-0 m-0")
}
\arguments{
\item{...}{html attributes and inner html}

\item{class}{[character]  Must have an exact length of 1.  Defaults to 'col-md-9 ml-sm-auto col-lg-10 p-0 m-0'}
}
\value{
\code{main}: html
}
\description{
Create a div for main content.
}
