% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_menu_bar.R
\name{main_menu_bar}
\alias{main_menu_bar}
\title{Create a main menu bar [File,Edit,Code,View,etc].}
\usage{
main_menu_bar(..., class = "")
}
\arguments{
\item{...}{should be named lists.  name represent the menu header and contents the dropdown options. Can be used recursively for multiple dropdown options.  See example.}

\item{class}{[character]  Must have an exact length of 1.  Defaults to ''}
}
\value{
\code{main_menu_bar}: [HTML]
}
\description{
Create a main menu bar [File,Edit,Code,View,etc].
}
\examples{
 main_menu_bar(File = list('New', 'Open', 'Save', list(l2 = list('one',
 'two', 'three', list(l3 = list('one', 'two', 'three'))))),
 Data = .('New data source', 'Paste'), View = .('Graph',
 'Dashboard'))
}
