% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialIconGallery.R
\name{materialIconGallery}
\alias{materialIconGallery}
\title{View the material design icon library.}
\usage{
materialIconGallery()
}
\value{
\code{materialIconGallery}: [NULL]
}
\description{
View the material design icon library.
}
\examples{
 materialIconGallery() 
}
