% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_dropdown.R
\name{ml_dropdown}
\alias{ml_dropdown}
\title{Create a multi level drop down.}
\usage{
ml_dropdown(..., class = "", is_submenu = FALSE)
}
\arguments{
\item{class}{[character]  Must have an exact length of 1.  Defaults to ''}

\item{is_submenu}{[logical]  Defaults to FALSE}
}
\value{
\code{ml_dropdown}: [HTML]
}
\description{
Create a multi level drop down.
}
\examples{
 header = 'Drop Down'
 ml_dropdown(class = 'multi-level-dropdown', dd_header(header,
 id = label2id(header), class = 'menu-text', carat = FALSE),
 dd_menu(id = label2id(header), class = 'shadow', dd_item(id = 'option1',
 label = 'Option 1'), dd_item(id = 'option2', label = 'Option 2'),
 dd_item(id = 'option3', label = 'Option 3')))
}
