% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_item.R
\name{mm_item}
\alias{mm_item}
\title{Create a multi-level dropdown menu item.}
\usage{
mm_item(
  items,
  header,
  class = "menu-item",
  carat = FALSE,
  submenu = FALSE,
  id = NULL
)
}
\arguments{
\item{items}{[list]}

\item{header}{[character]  Must have an exact length of 1.}

\item{class}{[character]  Must have an exact length of 1.  Defaults to 'menu-item'}

\item{carat}{[logical]  Must have an exact length of 1.  Defaults to FALSE}

\item{submenu}{[logical]  Must have an exact length of 1.  Defaults to FALSE}
}
\value{
\code{mm_item}: [HTML]
}
\description{
Create a multi-level dropdown menu item.
}
\examples{
 mm_item(item = list('New', 'Open', 'Save', list(l2 = list('one',
 'two', 'three'))), header = 'File', carat = FALSE, submenu = FALSE)
}
