% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav_brand.R
\name{nav_brand}
\alias{nav_brand}
\title{Create a BS4 navbar title.}
\usage{
nav_brand(...)
}
\arguments{
\item{...}{html attributes and inner html}
}
\value{
\code{nav_brand}: HTML
}
\description{
For use inside a fixed top nav
}
