% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav_fixed_top.R
\name{nav_fixed_top}
\alias{nav_fixed_top}
\title{Create a BS4 fixed top navbar.}
\usage{
nav_fixed_top(...)
}
\arguments{
\item{...}{html attributes and inner html}
}
\value{
\code{nav_fixed_top}: HTML
}
\description{
Create a BS4 fixed top navbar.
}
