% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav_toggler.R
\name{nav_toggler}
\alias{nav_toggler}
\title{Create a navbar toggler button.}
\usage{
nav_toggler(
  ...,
  class = NULL,
  data_target,
  expanded = FALSE,
  icon = "navbar-toggler-icon"
)
}
\arguments{
\item{class}{[character]  Must have an exact length of 1.  Defaults to NULL}

\item{data_target}{[character] the id of the targat nav. Must start with "#" and Must have an exact length of 1.}

\item{expanded}{[logical]  Must have an exact length of 1.  Defaults to FALSE}

\item{icon}{[character]  Must have an exact length of 1.  Defaults to 'navbar-toggler-icon'}
}
\value{
\code{nav_toggler}: NULL
}
\description{
Create a navbar toggler button.
}
