% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navItem.R
\name{navItem}
\alias{navItem}
\title{Create a navigation item.}
\usage{
navItem(..., tab_id, class = NULL, active = FALSE)
}
\arguments{
\item{...}{Inner HTML.  Usually a text label or icon.}

\item{tab_id}{[character] This links to the content area. Must be unique. Must have an exact length of 1.}

\item{class}{[character]  Must have an exact length of 1.  NULL is ok.  Defaults to NULL}

\item{active}{[logical]  Must have an exact length of 1.  Defaults to FALSE}
}
\value{
\code{navItem}: html
}
\description{
For use inside navList().
}
