% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navList.R
\name{navList}
\alias{navList}
\title{Create a navigation list.
Should be used in conjuction with navListContent}
\usage{
navList(..., nav_id, type = "tabs", class = NULL, format = NULL)
}
\arguments{
\item{nav_id}{[character] The navigation id. Used for name spacing.  This id should match the id used in navListContent(). Must have an exact length of 1.}

\item{type}{[choice("tabs"|"pills")]  Defaults to 'tabs'}

\item{class}{[character]  Must have an exact length of 1.  NULL is ok.  Defaults to NULL}

\item{format}{[choice("column"|"fill"|"justified")]  NULL is ok.  Defaults to NULL}
}
\value{
\code{navList}: html
}
\description{
Create a navigation list.
Should be used in conjuction with navListContent
}
\examples{
 navList(nav_id = 'data', class = 'main-nav', type = 'pills',
 format = 'justified', navItem(tab_id = 'one', 'One',
 active = TRUE), navItem(tab_id = 'two', 'Two'))
}
