% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navListContent.R
\name{navListContent}
\alias{navListContent}
\title{Create nav list content link to a navList.}
\usage{
navListContent(..., nav_id, class = "")
}
\arguments{
\item{nav_id}{[character]  Must have an exact length of 1.  NULL is ok.}

\item{class}{[character]  Must have an exact length of 1.  NULL is ok.  Defaults to ''}
}
\value{
\code{navListContent}: html
}
\description{
Should be used in conjunction with navList().
}
\examples{
 tagList(navList(nav_id = 'data', class = 'main-nav', type = 'pills',
 format = 'justified', navItem(tab_id = 'one', 'One',
 active = TRUE), navItem(tab_id = 'two', 'Two')),
 navListContent(nav_id = 'data', div(id = 'one', 'one'),
 div(id = 'two', 'two')))
}
