% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_shiny_app.R
\name{new_shiny_app}
\alias{new_shiny_app}
\title{Create a new BS4 app from a template.}
\usage{
new_shiny_app(app_name, template = "dashboardTemplate")
}
\arguments{
\item{app_name}{[character]  Must have an exact length of 1.}

\item{template}{[character]  Must have an exact length of 1.  Defaults to 'dashboardTemplate'}
}
\value{
\code{new_shiny_app}: NULL
}
\description{
Create a new BS4 app from a template.
}
