% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newShinyInputBinding.R
\name{newShinyInputBinding}
\alias{newShinyInputBinding}
\title{Create a new shiny input dependency.}
\usage{
newShinyInputBinding(name)
}
\arguments{
\item{name}{[string]}
}
\value{
\code{newShinyInputBinding}: NULL
}
\description{
Create a new shiny input dependency.
}
