% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picker_page_btn.R
\name{picker_page_btn}
\alias{picker_page_btn}
\title{Create a button to page a select picker.}
\usage{
picker_page_btn(
  inputId,
  pickerId,
  class = "btn",
  type = c("page-up", "page-down"),
  icon = "",
  icon_size = c("md"),
  ...
)
}
\arguments{
\item{inputId}{[string]}

\item{pickerId}{[string]}

\item{class}{[character]  Defaults to 'btn'}

\item{type}{[choice]  Possible values: c('c', 'page-up', 'page-down').  Defaults to c('page-up', 'page-down')}

\item{icon}{[string]  Defaults to ''}

\item{icon_size}{[choice]  Possible values: c( 'xs', 'sm', 'md', 'lg', 'xl').  Defaults to c('md')}
}
\value{
\code{picker_page_btn}: HTML
}
\description{
Create a button to page a select picker.
}
\examples{
 if (interactive()) {
 library('shiny')
 library(Schrock)
 library('ShinyReboot')
 library('bootstraplib')
 library('htmltools')
 bs_global_theme()
 ui <- fluidPage(bs_dependencies(theme = bs_global_get()),
 tags$head(reboot_Dependancies()), tags$h1('picker_page_btn'),
 div(class = 'btn-group', select_picker(inputId = 'font_size',
 choices = c('8px', '9px', '10px', '11px', '12px',
 '14px', '18px'), class = 'btn btn-outline-dark btn-sm rounded-0'),
 picker_page_btn(inputId = 'font_size_increase',
 pickerId = 'font_size', type = 'page-up', class = 'btn btn-outline-dark btn-sm rounded-0',
 icon = 'format-font-size-increase'), picker_page_btn(inputId = 'font_size_decrease',
 pickerId = 'font_size', type = 'page-down',
 class = 'btn btn-outline-dark btn-sm rounded-0',
 icon = 'format-font-size-decrease')), verbatimTextOutput(outputId = 'res'))
 server <- function(input, output, session) {
 output$res <- renderPrint(input$font_size)
 }
 shinyApp(ui, server)
 }
}
