% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pill_card.R
\name{pill_card}
\alias{pill_card}
\title{Create a pill card div.}
\usage{
pill_card(
  ...,
  inputId,
  name = NULL,
  icon,
  type = "column",
  sortable = TRUE,
  sort_ops = sortable_options()
)
}
\arguments{
\item{inputId}{\code{[string]}  Must have an exact length of 1.}

\item{name}{\code{[string]}  NULL is ok.  Defaults to NULL}

\item{icon}{\code{[string]}  Must have an exact length of 1.}

\item{type}{\code{[choice]}  Possible values: \code{c( 'row', 'column')}.  Defaults to 'column'}

\item{sortable}{\code{[logical]}  Must have an exact length of 1.  Defaults to TRUE}

\item{sort_ops}{\code{[class('sortable_options')]}  Should be the output of a sortable_options call. See \code{?sortable_options} for details.}
}
\value{
\code{pill_card}: [htmlTag(div)]
}
\description{
Create a pill card div.
}
\examples{
 pill_card(pill_item(id = c('sales', 'transactions', 'traffic',
 'basket size'), label = c('sales', 'transactions', 'traffic',
 'basket size')), inputId = 'one', name = 'Chart', icon = 'test')
}
