% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pill_folder.R
\name{pill_folder}
\alias{pill_folder}
\title{Create a pill folder.}
\usage{
pill_folder(folder_id, folder_name, pill_items, sortable_group)
}
\arguments{
\item{folder_id}{\code{[string]}}

\item{folder_name}{\code{[string]}}

\item{pill_items}{\code{[list]}}

\item{sortable_group}{\code{[string]}}
}
\value{
\code{pill_folder}: \code{[html]}
}
\description{
Create a pill folder.
}
\examples{
 pill_folder(folder_id = 'folder_id', folder_name = 'Sales Data',
 pill_items = pill_item(id = c('sales1', 'transactions1',
 'traffic1', 'basket_size1'), label = c('sales', 'transactions',
 'traffic', 'basket size'), data = list(nformat = c('currency',
 'integer', 'integer', NA), filter = c(TRUE, FALSE,
 NA, FALSE), edit = c(NA, 0, 0, 0), submenu_units = c(0,
 0, 0, NA))), sortable_group = 'measures')
}
