% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pill_item.R
\name{pill_item}
\alias{pill_item}
\title{Create pill items.}
\usage{
pill_item(
  id,
  label,
  type = "measure",
  data = list(),
  icon = "numeric",
  is_hierachy = FALSE
)
}
\arguments{
\item{id}{[character] the pill ids}

\item{label}{[character] the pill labels}

\item{type}{[character]  Must have an exact length of or equal to one of the following: [1,length(id)].  Defaults to 'measure'}
}
\value{
\code{pill_item}: [html]
}
\description{
Create pill items.
}
\examples{
 pill_card(pill_item(id = c('sales', 'transactions', 'traffic',
 'basket size'), label = c('sales', 'transactions', 'traffic',
 'basket size')), inputId = 'one', name = 'Chart', icon = 'test')
}
