% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pill_shelf.R
\name{pill_shelf}
\alias{pill_shelf}
\title{Create a variable pill shelf.}
\usage{
pill_shelf(ids, data, group_name, session = getDefaultReactiveDomain())
}
\arguments{
\item{ids}{\code{[list]}}

\item{data}{\code{[list]}}

\item{group_name}{\code{[string]}}

\item{session}{\code{[NULL]}  Defaults to \code{getDefaultReactiveDomain()}.}
}
\value{
\code{pill_shelf}: \code{[html]}
}
\description{
Create a variable pill shelf.
}
