% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radio_btn_group.R
\name{radio_btn_group}
\alias{radio_btn_group}
\title{Create BS4 radio button groups.}
\usage{
radio_btn_group(
  inputId,
  labels,
  selected = NULL,
  btn_class = "btn btn-sm",
  group_class = NULL
)
}
\arguments{
\item{selected}{[NULL]  Defaults to NULL}

\item{btn_class}{[character]  Defaults to 'btn btn-sm'}

\item{group_class}{[NULL]  Defaults to NULL}
}
\value{
\code{radio_btn_group}: HTML
}
\description{
Create BS4 radio button groups.
}
\examples{
 if (interactive()) {
 library(shiny)
 library(bootstraplib)
 library(exprTools)
 library(ShinyReboot)
 bs_global_theme()
 ui <- fluidPage(tags$head(bs_dependencies(theme = bs_global_get()),
 reboot_Dependancies()), tags$h1('radio_btn_group examples'),
 radio_btn_group(inputId = 'somevalue1', labels = c('A',
 'B', 'C'), selected = 'B', btn_class = 'btn btn-sm btn-primary'),
 verbatimTextOutput('value1'), )
 server <- function(input, output) {
 output$value1 <- renderPrint({
 input$somevalue1
 })
 }
 shinyApp(ui, server)
 }
}
