% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radio_slider_input.R
\name{radio_slider_input}
\alias{radio_slider_input}
\title{Create a radio slider input.}
\usage{
radio_slider_input(
  ...,
  inputId,
  label = NULL,
  units_label = NULL,
  choices,
  selected = NULL,
  class = "w-100"
)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{label}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}

\item{units_label}{\code{[NULL]}  Defaults to \code{NULL}.}

\item{choices}{\code{[atomic_vector]}}

\item{selected}{\code{[choice]}  Possible values: \code{choices}.  NULL is ok.  Defaults to \code{NULL}}

\item{class}{\code{[string]}  NULL is ok.  Defaults to \code{'w-100'}}
}
\value{
\code{radio_slider_input}: \code{[html]}
}
\description{
Create a radio slider input.
}
\examples{
 radio_slider_input(inputId = 'myid', choices = names(mtcars))
}
