% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_slider_input.R
\name{range_slider_input}
\alias{range_slider_input}
\title{Create a range slider input.}
\usage{
range_slider_input(
  ...,
  inputId,
  label = NULL,
  units_label = NULL,
  range,
  lower_val = NULL,
  upper_val = NULL,
  prefix = NULL,
  suffix = NULL,
  signif = 5,
  class = "w-100",
  variable_id = NULL
)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{label}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}

\item{units_label}{\code{[NULL]}  Defaults to \code{NULL}.}

\item{lower_val}{\code{[number]}  NULL is ok.  Defaults to \code{NULL}}

\item{upper_val}{\code{[number]}  NULL is ok.  Defaults to \code{NULL}}

\item{prefix}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}

\item{suffix}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}

\item{signif}{\code{[int]}  Defaults to \code{5}}

\item{class}{\code{[string]}  NULL is ok.  Defaults to \code{'w-100'}}

\item{min}{\code{[number]}}

\item{max}{\code{[number]}  Must be greater than \code{min}.}
}
\value{
\code{range_slider_input}: \code{[html]}
}
\description{
Create a range slider input.
}
\examples{
 range_slider_input(inputId = 'myid', label = NULL, min = 0,
 max = 100, lower_val = 0, upper_val = 100, prefix = NULL,
 suffix = '\%', signif = 4)
}
