% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reboot_Dependancies.R
\name{reboot_Dependancies}
\alias{reboot_Dependancies}
\title{Loads variable scripts from shiny applications.}
\usage{
reboot_Dependancies()
}
\value{
\code{reboot_Dependancies}: [list] HTML
}
\description{
Loads variable scripts from shiny applications.
}
