% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_class.R
\name{remove_class}
\alias{remove_class}
\title{Remove the class of a shiny tag.}
\usage{
remove_class(x, class)
}
\arguments{
\item{x}{[class]}

\item{class}{[character]}
}
\value{
\code{remove_class}: shiny tag
}
\description{
Remove the class of a shiny tag.
}
\examples{
 x <- div(class = 'class1 class2 active show')
 remove_class(x) <- 'active'
 x
 remove_class(x) <- 'class1 active'
 x
 x <- div(class = 'class1 class2 active show')
 remove_class(x) <- c('class1', 'active')
 x
}
