% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_mdi_icons.R
\name{search_mdi_icons}
\alias{search_mdi_icons}
\title{Search the material icon list.}
\usage{
search_mdi_icons(pattern)
}
\arguments{
\item{pattern}{[character]  Must have an exact length of 1.}
}
\value{
\code{search_mdi_icons}: [character]
}
\description{
Search the material icon list.
}
\examples{
 search_mdi_icons('database')
}
