% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_group.R
\name{select_group}
\alias{select_group}
\title{Create a button group that can include select dropdowns.}
\usage{
select_group(..., group_class = NULL, env = caller_env())
}
\arguments{
\item{group_class}{[NULL]  Defaults to NULL}
}
\value{
\code{select_group}: HTML
}
\description{
Create a button group that can include select dropdowns.
}
\examples{
 if (interactive()) {
 library(shiny)
 library(Schrock)
 library(ShinyReboot)
 library(bootstraplib)
 library(htmltools)
 bs_global_theme()
 ui <- fluidPage(bs_dependencies(theme = bs_global_get()),
 tags$head(reboot_Dependancies()), tags$h1('picker_page_btn'),
 select_group(select_picker(inputId = 'ace_theme',
 choices = getAceThemes(), selected = 'pastel_on_dark'),
 select_picker(inputId = 'font_size', choices = c('8px',
 '9px', '10px', '11px', '12px', '14px', '18px'),
 ), picker_page_btn(inputId = 'font_size_increase',
 pickerId = 'font_size', type = 'page-up', icon = 'format-font-size-increase'),
 picker_page_btn(inputId = 'font_size_decrease',
 pickerId = 'font_size', type = 'page-down',
 icon = 'format-font-size-decrease'), class = 'btn btn-outline-dark btn-sm'),
 verbatimTextOutput(outputId = 'res'), verbatimTextOutput(outputId = 'res2'))
 server <- function(input, output, session) {
 output$res <- renderPrint(input$font_size)
 output$res2 <- renderPrint(input$ace_theme)
 }
 shinyApp(ui, server)
 }
}
