% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendCustomMessage.R
\name{sendCustomMessage}
\alias{sendCustomMessage}
\title{Send a shiny message to the browser.}
\usage{
sendCustomMessage(id, message, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{\code{[string]}}

\item{session}{\code{[NULL]}  Defaults to \code{getDefaultReactiveDomain()}.}
}
\value{
\code{sendCustomMessage}: \code{[invisible(NULL)]}
}
\description{
Send a shiny message to the browser.
}
