% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_ns.R
\name{session_ns}
\alias{session_ns}
\title{Get a session namespace}
\usage{
session_ns(session = getDefaultReactiveDomain())
}
\arguments{
\item{session}{the shiny session.}
}
\value{
the session's namespace function
}
\description{
Get a session namespace
}
