% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ace_tokens.R
\name{set_ace_tokens}
\alias{set_ace_tokens}
\title{Set highlighting tokens for an ace editor.}
\usage{
set_ace_tokens(
  inputId,
  tokens,
  as_is = FALSE,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{tokens}{\code{[list]}}

\item{as_is}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{FALSE}}

\item{session}{\code{[NULL]}  Defaults to \code{getDefaultReactiveDomain()}.}
}
\value{
\code{set_ace_tokens}: \code{[invisible(NULL)]}
}
\description{
Set highlighting tokens for an ace editor.
}
