% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_html_data.R
\name{set_html_data}
\alias{set_html_data}
\title{Set data on an HTML element/s.}
\usage{
set_html_data(selector, data, trigger_event = NULL)
}
\arguments{
\item{selector}{\code{[string]}}

\item{data}{\code{[named]}}

\item{trigger_event}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}
}
\value{
\code{set_html_data}: \code{[invisible(NULL)]}
}
\description{
Set data on an HTML element/s.
}
