% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInputValue.R
\name{setInputValue}
\alias{setInputValue}
\title{Set a shiny input value from the server side.}
\usage{
setInputValue(inputId, value, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{[string]}

\item{value}{the new value of the input}

\item{session}{[NULL]  Defaults to shiny::getDefaultReactiveDomain()}
}
\value{
\code{setInputValue}: NULL
}
\description{
Set a shiny input value from the server side.
}
\examples{
 if (interactive()) {
 library('shiny')
 library('ShinyReboot')
 ui <- fluidPage(tagList(html_dependency_setShinyInput()),
 tags$h1('Set a shiny input value'), verbatimTextOutput(outputId = 'res'))
 server <- function(input, output, session) {
 output$res <- renderPrint(input$myColor)
 observe({
 setInputValue('myColor', '#ffffff')
 })
 }
 shinyApp(ui, server)
 }
}
