% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_draggable_model.R
\name{show_draggable_model}
\alias{show_draggable_model}
\title{show or hide a draggable model.}
\usage{
show_draggable_model(
  inputId = NULL,
  top = NULL,
  left = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{\code{[string]}  Defaults to \code{NULL}}

\item{top}{\code{[number]}  NULL is ok.  Defaults to \code{NULL}}

\item{left}{\code{[number]}  NULL is ok.  Defaults to \code{NULL}}

\item{session}{\code{[NULL]}  Defaults to \code{getDefaultReactiveDomain()}.}
}
\value{
\code{show_draggable_model}: invisible(NULL)
}
\description{
show or hide a draggable model.
}
