% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar_left.R
\name{sidebar_left}
\alias{sidebar_left}
\title{Create a left sidebar.}
\usage{
sidebar_left(..., id)
}
\arguments{
\item{...}{html attributes and inner html}

\item{id}{[character]  Must have an exact length of 1.}
}
\value{
\code{sidebar_left}: html
}
\description{
Create a left sidebar.
}
