% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar_sticky.R
\name{sidebar_sticky}
\alias{sidebar_sticky}
\title{Create a sticky sidebar.}
\usage{
sidebar_sticky(...)
}
\arguments{
\item{...}{html attributes and inner html}
}
\value{
\code{sidebar_sticky}: html
}
\description{
Should be used inside sidebar_left/right.
}
