% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortable_insert.R
\name{sortable_insert}
\alias{sortable_insert}
\title{insert html into a sortable div.}
\usage{
sortable_insert(
  ...,
  inputId,
  where = "start",
  data_id = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{where}{\code{[choice]}  Possible values: \code{c('start', 'end', 'before', 'after')}.  Defaults to \code{'start'}}

\item{data_id}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}

\item{session}{\code{[NULL]}  Defaults to \code{getDefaultReactiveDomain()}.}
}
\value{
\code{sortable_insert}: \code{[invisible(NULL)]}
}
\description{
insert html into a sortable div.
}
