% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortable_remove.R
\name{sortable_remove}
\alias{sortable_remove}
\title{insert html into a sortable div.}
\usage{
sortable_remove(inputId, data_id, session = getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{\code{[string]} The sortable inputId}

\item{data_id}{\code{[character]}  The a vector of item data ids to remove.}

\item{session}{\code{[NULL]}  Defaults to \code{getDefaultReactiveDomain()}.}
}
\value{
\code{sortable_insert}: \code{[invisible(NULL)]}
}
\description{
insert html into a sortable div.
}
