% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortableDiv.R
\name{sortableDiv}
\alias{sortableDiv}
\title{Create a sortable div.}
\usage{
sortableDiv(..., inputId, options = sortable_options())
}
\arguments{
\item{...}{sortable HTML elements.  Each element should have a unqiue `data-id` attribute.}

\item{inputId}{\code{[string]}}

\item{options}{\code{[class('sortable_options')]}  Should be the output of a sortable_options call. See \code{?sortable_options} for details.}
}
\value{
\code{sortableDiv}: HTML
}
\description{
Create a sortable div.
}
\examples{
 sortableDiv(class = 'p-1', inputId = 'sortablejs1', div(class = 'm-1 bg-secondary',
 `data-id` = 'one', 'one'), div(class = 'm-1 bg-secondary',
 `data-id` = 'two', 'two'), div(class = 'm-1 bg-secondary',
 `data-id` = 'three', 'three'), options = sortable_options(name = 'things'))
}
