% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_dropdown.R
\name{standard_dropdown}
\alias{standard_dropdown}
\title{Create a standard dropdown.}
\usage{
standard_dropdown(
  ...,
  inputId,
  target,
  handle_class = NULL,
  class = NULL,
  width = "auto"
)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{handle_class}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}

\item{class}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}

\item{target_class}{\code{[string]}}
}
\value{
\code{standard_dropdown}: \code{[html]}
}
\description{
Create a standard dropdown.
}
