% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_checkbox_group_input.R
\name{update_checkbox_group_input}
\alias{update_checkbox_group_input}
\title{Update a checkbox_grou_input.}
\usage{
update_checkbox_group_input(
  inputId,
  label = NULL,
  choices = NULL,
  selected = NULL,
  data_attrs = NULL
)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{label}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}

\item{choices}{\code{[atomic_vector]}  Defaults to \code{NULL}}

\item{selected}{\code{[choice]}  Possible values: \code{choices}.  NULL is ok.  Defaults to \code{NULL}}

\item{data_attrs}{\code{[list]}  NULL is ok.  Defaults to \code{NULL}}
}
\value{
\code{update_checkbox_group_input}: \code{[invisible(NULL)]}
}
\description{
Update a checkbox_grou_input.
}
