% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_default_color_picker.R
\name{update_default_color_picker}
\alias{update_default_color_picker}
\title{Update a default color picker input.}
\usage{
update_default_color_picker(
  inputId,
  value,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{value}{\code{[string]}  All non-missing elements must comply to regex pattern \code{'''^#'''}.}

\item{session}{\code{[NULL]}  Defaults to \code{getDefaultReactiveDomain()}.}
}
\value{
\code{update_default_color_picker}: \code{[invisible(NULL)]}
}
\description{
Update a default color picker input.
}
