% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_discrete_variable_pal.R
\name{update_discrete_variable_pal}
\alias{update_discrete_variable_pal}
\title{update a discrete variable palette input.}
\usage{
update_discrete_variable_pal(
  inputId,
  var_names,
  var_ids = NULL,
  colors = NULL,
  palette = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{var_names}{\code{[character]}}

\item{var_ids}{\code{[character]}  NULL is ok.  Defaults to \code{NULL}}

\item{colors}{\code{[character]}  NULL is ok.  Must have an exact length of \code{length(var_names)}.  Defaults to \code{NULL}}

\item{session}{\code{[NULL]}  Defaults to \code{getDefaultReactiveDomain()}.}
}
\value{
\code{update_discrete_variable_pal}: \code{[invisible(NULL)]}
}
\description{
update a discrete variable palette input.
}
