% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_editable.R
\name{update_editable}
\alias{update_editable}
\title{Update an editable class.}
\usage{
update_editable(
  inputId,
  data_id,
  value = NULL,
  trigger_edit = FALSE,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{data_id}{\code{[string]}}

\item{value}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}

\item{trigger_edit}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{FALSE}}

\item{session}{\code{[NULL]}  Defaults to \code{getDefaultReactiveDomain()}.}
}
\value{
\code{update_editable}: \code{[invisible(NULL)]}
}
\description{
Update an editable class.
}
