% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_range_slider_input.R
\name{update_range_slider_input}
\alias{update_range_slider_input}
\title{Update a range slider input.}
\usage{
update_range_slider_input(
  ...,
  inputId,
  lower_val = NULL,
  upper_val = NULL,
  range = NULL,
  label = NULL,
  signif = NULL,
  prefix = NULL,
  suffix = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{lower_val}{\code{[number]}  NULL is ok.  Defaults to \code{NULL}}

\item{upper_val}{\code{[number]}  NULL is ok.  Defaults to \code{NULL}}

\item{range}{\code{[numeric]}  Must have an exact length of \code{2}.  NULL is ok.  Defaults to \code{NULL}}

\item{label}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}

\item{signif}{\code{[int]}  NULL is ok.  Defaults to \code{NULL}}

\item{prefix}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}

\item{suffix}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}

\item{session}{\code{[NULL]}  Defaults to \code{getDefaultReactiveDomain()}.}
}
\value{
\code{update_range_slider_input}: \code{[invisible(NULL)]}
}
\description{
Update a range slider input.
}
\examples{
 update_range_slider_input(inputId = 'myid', lower_val = 4,
 upper_val = 5, range = c(0, 10), label = 'new_label',
 signif = 4, prefix = '$')
}
