% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateInput.R
\name{updateInput}
\alias{updateInput}
\title{Update a shiny input.}
\usage{
updateInput(inputId, value, session = getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{\code{[string]}}

\item{session}{\code{[NULL]}  Defaults to \code{getDefaultReactiveDomain()}.}
}
\value{
\code{updateInput}: \code{[invisible(NULL)]}
}
\description{
Update a shiny input.
}
