% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateInputTemplate.R
\name{updateInputTemplate}
\alias{updateInputTemplate}
\title{Generate a template function for updating a custom shiny input.}
\usage{
updateInputTemplate(inputName)
}
\arguments{
\item{inputName}{[string]}
}
\value{
\code{updateInputTemplate}: [character(1)]
}
\description{
Generate a template function for updating a custom shiny input.
}
\examples{
 updateInputTemplate('CustomInput')
}
