% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSortableDiv.R
\name{updateSortableDiv}
\alias{updateSortableDiv}
\title{Update a sortableDiv}
\usage{
updateSortableDiv(
  inputId,
  ...,
  content = NULL,
  order = NULL,
  append = NULL,
  clear = FALSE,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{the id of the sortable}

\item{...}{sortable options to update}

\item{content}{Shiny tag or list of shiny tags.  Replaces the content of the sortable. Each element should have a data-id attribute.}

\item{order}{character vector of data ids.  Updates the order of the list.}

\item{append}{Shiny tag or list of shiny tags. Appends the content of the sortable. Each element should have a data-id attribute.}

\item{clear}{[Logical(1)] If TRUE, the content of the sortable will be removed.}
}
\description{
Update a sortableDiv
}
