% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_arrange.R
\name{v_arrange}
\alias{v_arrange}
\title{Arrange html content.}
\usage{
v_arrange(..., justify = "around", align = "center")
}
\arguments{
\item{justify}{\code{[choice]}  Possible values: \code{c('start', 'end', 'center', 'between', 'around')}.  Defaults to \code{'around'}}

\item{align}{\code{[choice]}  Possible values: \code{c('start', 'end', 'center', 'baseline', 'stretch')}.  Defaults to \code{'center'}}
}
\value{
\code{v_arrange}: \code{[html]}
}
\description{
Arrange html content.
}
\examples{
 v_arrange(justify = 'around', align = 'center')
 v_arrange(justify = 'start', align = 'start')
}
