% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_mdi_icons.R
\name{view_mdi_icons}
\alias{view_mdi_icons}
\title{Search and view the material icon list.}
\usage{
view_mdi_icons(icon_names, size = 25)
}
\arguments{
\item{icon_names}{[character]  Must have an exact length of 1.}

\item{size}{[number]  Icon size in px.}
}
\value{
\code{view_mdi_icons}: [NULL]
}
\description{
Search and view the material icon list.
}
\examples{
 view_mdi_icons('database')
}
