% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorize_user.R
\name{authorize_user}
\alias{authorize_user}
\title{Authorize app access.}
\usage{
authorize_user(id, user = NULL, pw = NULL, path = "udb.rds", app_ui = NULL)
}
\arguments{
\item{id}{\code{[string]}}

\item{user}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}

\item{pw}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}

\item{path}{\code{[file]}  Defaults to \code{'udb.rds'}}

\item{app_ui}{\code{[function]}  NULL is ok.  Defaults to \code{NULL}}
}
\value{
\code{authorize_user}: \code{[character(1)]}
}
\description{
Authorize app access.
}
\examples{
 if (interactive()) {
 library(shiny)
 library(ShinyReboot)
 library(bslib)
 bslib::bs_global_theme()
 ui <- fluidPage(tags$head(bslib::bs_theme_dependencies(theme = bs_global_get())),
 titlePanel('authorize module example'), uiOutput('loggedin'))
 server <- function(input, output, session) {
 out <- authorize_user(id = 'mod_id')
 output$loggedin <- renderUI({
 req(out())
 fluidRow(h3(glue('Logged in as {out()}')))
 })
 }
 shinyApp(ui, server)
 }
}
