% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dropdown.R
\name{data_dropdown}
\alias{data_dropdown}
\title{Create a data dropdown.}
\usage{
data_dropdown(..., inputId, target_class, handle_class = NULL, class = NULL)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{target_class}{\code{[string]}}

\item{handle_class}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}

\item{class}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}
}
\value{
\code{data_dropdown}: \code{[html]}
}
\description{
Create a data dropdown.
}
\examples{
 data_dropdown(dropdown_button(inputId = 'edit', label = 'Edit...'),
 dropdown_checkbox(inputId = 'filter', value = FALSE,
 label = 'Show Filter'), dropdown_radio_group(inputId = 'radio_nformat',
 options = c('currency', 'percentage', 'float', 'integer'),
 labels = c('currency', 'percentage', 'float', 'integer')),
 dropdown_submenu(label = 'Units', dropdown_radio_group(inputId = 'radio_units',
 labels = c('duration', 'distance', 'mass', 'speed'),
 options = c('duration', 'distance', 'mass', 'speed'))),
 inputId = 'myid', handle_class = 'dropdown-handle', target_class = 'pill-item')
}
