% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draggable_model_btn.R
\name{draggable_model_btn}
\alias{draggable_model_btn}
\title{Create a button to show or hide a draggable model.}
\usage{
draggable_model_btn(..., inputId, toggleId, outerTag = "button")
}
\arguments{
\item{inputId}{\code{[string]}}

\item{toggleId}{\code{[string]}}

\item{outerTag}{\code{[string]}  Defaults to \code{'button'}}
}
\value{
\code{draggable_model_btn}: \code{[html]}
}
\description{
Create a button to show or hide a draggable model.
}
\examples{
 draggable_model_btn(inputId = 'drag_btn', toggleId = 'drag',
 'Show')
 draggable_model_btn(inputId = 'drag_btn', toggleId = 'drag',
 'Show', outerTag = 'div')
}
