% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_mdi.R
\name{icon_mdi}
\alias{icon_mdi}
\title{Generate a materical icon.}
\usage{
icon_mdi(name, class = "")
}
\arguments{
\item{name}{[character]  Must have an exact length of 1.}

\item{class}{[character]  Must have an exact length of 1.  Defaults to ''}
}
\value{
\code{icon_mdi}: [htmlTag(span)]
}
\description{
Generate a materical icon.
}
\examples{
 icon_mdi('undo')
}
