% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_toggler.R
\name{icon_toggler}
\alias{icon_toggler}
\title{Create an icon toggler.}
\usage{
icon_toggler(inputId, value = FALSE, icon_false, icon_true, ...)
}
\arguments{
\item{inputId}{\code{[string]}}

\item{value}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{FALSE}}

\item{icon_false}{\code{[string]}}

\item{icon_true}{\code{[string]}}
}
\value{
\code{icon_toggler}: \code{[html]}
}
\description{
Create an icon toggler.
}
\examples{
 icon_toggler(inputId = 'myid', value = FALSE, icon_false = 'folder-outline',
 icon_true = 'folder-open-outline')
 icon_toggler(inputId = 'myid', value = TRUE, icon_false = 'folder-outline',
 icon_true = 'folder-open-outline', class = 'icon-sm')
}
