% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pill_card.R
\name{pill_card}
\alias{pill_card}
\title{Create a pill card div.}
\usage{
pill_card(..., inputId, name = NULL, icon, type = "column", sortable = TRUE)
}
\arguments{
\item{inputId}{[character]  Must have an exact length of 1.}

\item{name}{[character]  NULL is ok.  Defaults to NULL}

\item{icon}{[character]  Must have an exact length of 1.}

\item{type}{[choice]  Possible values: c('c', 'row', 'column').  Defaults to 'column'}

\item{sortable}{[logical]  Must have an exact length of 1.  Defaults to TRUE}
}
\value{
\code{pill_card}: [htmlTag(div)]
}
\description{
Create a pill card div.
}
\examples{
 pill_card(pill_item(id = c('sales', 'transactions', 'traffic',
 'basket size'), label = c('sales', 'transactions', 'traffic',
 'basket size')), inputId = 'one', name = 'Chart', icon = 'test')
}
