% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ace_tokens.R
\name{set_ace_tokens}
\alias{set_ace_tokens}
\title{Set highlighting tokens for an ace editor.}
\usage{
set_ace_tokens(
  inputId,
  tokens,
  as_is = FALSE,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{\code{[string]} ace editor input id.}

\item{tokens}{\code{[list]} an unnamed list with structure:\code{list(list(token='tokenname',regex='pattern'),list(token='tokenname',regex='pattern'),...))}}

\item{as_is}{\code{[logical]}  If TRUE, use the ID as-is even when inside a module (instead of adding the namespace prefix to the ID). Must have an exact length of \code{1}.  Defaults to \code{FALSE}}

\item{session}{\code{[NULL]}  Defaults to \code{getDefaultReactiveDomain()}.}
}
\value{
\code{set_ace_tokens}: \code{[invisible(NULL)]}
}
\description{
Set highlighting tokens for an ace editor.
}
\examples{
set_ace_tokens(inputId='ace_id',tokens=list(list(token='token_name1',regex='one|two|three'),list(token='token_name1',regex='a|b|c')))
 if(interactive()){
 library("shiny")
 library("ShinyReboot")
 library("bslib")
 bs_global_theme()
 ui <- fluidPage(
 bs_theme_dependencies(theme = bs_global_get()),

 tags$h1("Ace Editor: Custom sythax highligher"),
 ace_editor('editor',mode='r',value='ifelse(cut=="ideal",max(price),min(price))'),
 actionButton('set_tokens','set custom tokens')
 )

 server = function(input, output, session) {

 observeEvent(input$set_tokens,{
 req(input$set_tokens)
 library(ggplot2)
 diamonds=data.table(diamonds)
 fn=function(x){is(x,'integer')|is(x,'numeric')}

 c(nums,identity)\%<-\%
 split_vec(names(diamonds),unlist(sapply(diamonds,fn )))

 functions = paste0(c('sum', 'min', 'max'), followed_by('\\\\('))
 tokens=c(
 map2(
 c('measure', 'dimension'),
 list(nums, identity),
 create_ace_token,
 escape = TRUE
 ),list(create_ace_token('function',functions,escape = FALSE)))

 set_ace_tokens('editor',tokens = tokens)
 })

 }
 shinyApp(ui = ui, server = server)
 }
}
